requirements.txt
python-dotenv
openai
perplexityai
